backend/requirements.txt
django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=25.2
python-dotenv>=1.2
psycopg2-binary>=2.9
gunicorn>=23.0
whitenoise>=6.6
dj-database-url>=3.0
google-genai>=1.0
requests>=2.31
Pillow>=10.0
markdown>=3.5
